Risolto un bug che faceva perdere un sacco di memoria e impediva di leggere correttamente i messaggi arrivati

+++ src/Client/messages/FilterAdd.java
@@ -24,7 +24,7 @@ public class FilterAdd extends Message{
 
     @Override
     public void read(LittleEndianInputStream leis) throws IOException {
-        data = new byte [getLength()];
+        data = new byte [(int) getLength()];
         leis.read(data);
     }
